+++ common/config/rush/repo-state.json
@@ -1,5 +1,5 @@
 // DO NOT MODIFY THIS FILE MANUALLY BUT DO COMMIT IT. It is generated and used by Rush.
 {
-  "pnpmShrinkwrapHash": "6ac773d01b5c279a78fb910fcf3763fa4ea0c586",
+  "pnpmShrinkwrapHash": "8d43354d253d8dc24457dda4d536ebaa33775088",
   "preferredVersionsHash": "bf21a9e8fbc5a3846fb05b4fa0859e0917b2202f"
 }
