+++ packages/apps/kadena-docs/package.json
@@ -37,6 +37,7 @@
     "@mdx-js/react": "~2.3.0",
     "@next/mdx": "~13.3.4",
     "acorn": "~8.8.2",
+    "date-fns": "~2.30.0",
     "js-yaml": "~4.1.0",
     "mdast-util-to-string": "~3.2.0",
     "next": "13.0.6",
