--- packages/apps/kadena-docs/src/types/ApiResponse.ts
@@ -1,10 +1,0 @@
-export interface IResponse<T> {
-  status: number;
-  message: string;
-  body?: T;
-}
-
-export interface ITopDoc {
-  label: string;
-  url: string;
-}
