+++ packages/apps/kadena-docs/tsconfig.json
@@ -6,7 +6,7 @@
       "@/*": ["./src/*"]
     },
     "target": "ES2020",
-    "allowJs": false,
+    "allowJs": true,
     "skipLibCheck": true,
     "noEmit": true,
     "incremental": true,
